Refactor CubeCompare into own page.
